Remove non-generic sliders
