+++ src/view/components/toolbar/Gesture.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
-import { CONSTANTS } from "../../../constants";
-import { Dropdown } from "../../Dropdown";
-import SensorButton from "../SensorButton";
+import { CONSTANTS } from "../../constants";
+import { Dropdown } from "../Dropdown";
+import SensorButton from "./SensorButton";
 
 const GESTURE_BUTTON_MESSAGE = "Send Gesture";
 interface IProps {
